database modal

+++ database/migrations/2022_06_23_093853_create_mode_of_payments_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('mode_of_payments', function (Blueprint $table) {
             $table->id();
-            $table->int('3_months');
-            $table->int('6_months');
+            $table->tinyInteger('3_months');
+            $table->tinyInteger('6_months');
             $table->integer('amount_of_initial_disposal');
             $table->timestamps();
         });
